Add InputsLoader
