+++ src/main.ts
@@ -1,4 +1,3 @@
-import {getInput} from '@actions/core';
 import {getOctokit, context} from '@actions/github';
 import {GitHub} from '@actions/github/lib/utils';
 
@@ -6,18 +5,11 @@ import {ActionProcessor} from './classes/action-processor';
 import {CommentGenerator} from './classes/comment-generator';
 import {ConfigParser} from './classes/config-parser';
 import {ContextParser} from './classes/context-parser';
-import {Inputs} from './interfaces';
-
-function getInputs(): Inputs {
-  return {
-    GithubToken: getInput('github_token'),
-    ConfigFilePath: getInput('config_file')
-  };
-}
+import {Inputs, InputsLoader} from './classes/inputs-loader';
 
 export async function run(): Promise<void> {
   try {
-    const inputs: Inputs = getInputs();
+    const inputs: Inputs = new InputsLoader();
     const githubClient: InstanceType<typeof GitHub> = getOctokit(inputs.GithubToken);
     const contextParser: ContextParser = new ContextParser(inputs, context);
     const configParser: ConfigParser = new ConfigParser(contextParser.runContext);
